METAMAC-1457: [CODES] Órdenes. CAMBIO IMPORTANTE: órdenes en columnas de la tabla TB_M_CODES
git-svn-id: http://svn.arte-consultores.com/svn/arte/istac/metamac/metamac-srm/trunk/metamac-srm@44087 05f23f4b-1f0b-0410-ba1b-ad9509270e59

+++ metamac-srm-core/src/main/java/org/siemac/metamac/srm/core/code/domain/shared/CodeMetamacVisualisationResult.java
@@ -12,7 +12,7 @@ public class CodeMetamacVisualisationResult implements Serializable {
     private Long                           parentIdDatabase;
     private CodeMetamacVisualisationResult parent;
     private String                         name;
-    private Long                           order;
+    private Integer                        order;
 
     public CodeMetamacVisualisationResult() {
     }
@@ -65,11 +65,11 @@ public class CodeMetamacVisualisationResult implements Serializable {
         this.name = name;
     }
 
-    public Long getOrder() {
+    public Integer getOrder() {
         return order;
     }
 
-    public void setOrder(Long order) {
+    public void setOrder(Integer order) {
         this.order = order;
     }
